refactor(lessons): extraire LessonChapterSidebar de StudentLessonView (#28)
Tranche 2 (suite) de la décomposition de StudentLessonView.vue.

- LessonChapterSidebar.vue (<script setup>) : sidebar « Sommaire » (méta leçon
  + liste de navigation des chapitres). Props lesson/chapters/
  activeChapterIndex/completedChapters/collapsed ; émet select/toggle.
  getContentTypeLabel via utils/lessonContent. CSS scopée (+ responsive) copiée
  à l'identique.
- La vue utilise <LessonChapterSidebar> ; aside + CSS sidebar (~150 l. + responsive)
  retirés (1345 -> 1141 l.). getContentTypeLabel/isChapterCompleted restent au
  parent (encore utilisés : badge en-tête + bouton bas).

Cumul StudentLessonView : 1547 -> 1141 l. (-26 %).

Parité visuelle : CSS à l'identique + build vert, QA manuelle recommandée.

Tests 261/261, contract 28/28, build OK.
